Crate httpre

Structs

Get

feed get with an url and we will retrieve the response for you

KvPair

命令行中的 key = value 可以通过 parse_kv_pair 解析成 KvPair 结构

Opts

A native httpie implementation with Rust, can you imagine how easy it is?

Post

feed post with an url and optional key=value pairs. We will post the data as JSON, and retrieve the response for you

Enums

SubCommand

Functions

get

get_content_type

将服务器返回的 content-type 类型解析成 Mime 类型

main

parse_kv_pair

因为我们为 KvPair 实现了 FromStr, 这里可以直接 s.parse() 得到 KvPair

parse_url

post

print_body

打印服务器返回的 HTTP body

print_body_sytect

使用 syntect 语法高亮继续完善我们的 HTTPie，分别parse html, js, json进行高亮

print_headers

print_resp

打印整个响应

print_status